Retain the hardware-tested 3DS release and simplify community documentation
Publish the pre-rewrite runtime and its supporting source, keeping the experimental renderer out of the retained release. The restored release rebuilds byte-for-byte to the hardware-tested executable and ELF; CIA code and container verification also pass.

Rewrite the project overview and setup guides for Melee and 3DS homebrew enthusiasts, with realistic performance expectations, FBI installation, build prerequisites, controls, and known limitations. Mark development complete for now and retire iteration notes from the current documentation. Preserve upstream source, history, and third-party notices.

+++ native-3ds/tests/audio_src_tests.c
@@ -18,7 +18,7 @@ int main(void){
     }
     int n=1;uint32_t phase=123;int16_t h[4]={0},out[96];
     mp_src_block(out,96,123456,&phase,h,1,ramp,&n);
-    for(unsigned i=0;i<96;++i)assert(out[i]==13*(i+1));
+    for(unsigned i=0;i<96;++i)assert(out[i]==(int)(13*(i+1)));
     assert(phase==123&&h[0]==13*93&&h[3]==13*96);
     printf("AX source conversion: %u fractional samples, block continuity and direct mode passed\n",tests);
 }
